requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3

# ── LangChain / LangGraph ─────────────────────────────────────────────────────
langgraph==0.2.56
langchain-core==0.3.25
langchain-community==0.3.12
langchain-google-genai==2.0.7

# ── Google Gemini SDK ─────────────────────────────────────────────────────────
# google-genai ≥ 1.9 fixes the aiohttp.ClientConnectorDNSError AttributeError
google-genai==1.9.0
google-generativeai==0.8.3

# ── aiohttp: must be ≥ 3.9 to match google-genai 1.9 expectations ─────────────
aiohttp==3.11.10

# ── LightRAG ──────────────────────────────────────────────────────────────────
lightrag-hku==1.3.6

# ── PDF parsing ───────────────────────────────────────────────────────────────
pypdf==5.1.0
pytesseract==0.3.13

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
numpy==1.26.4
